Отрефакторил код карты и формы
