+++ js/form.js
@@ -6,7 +6,15 @@ const form = document.querySelector('.ad-form');
 const mapFilters = document.querySelector('.map__filters');
 const fieldAddress = form.querySelector('#address');
 const DEGREE = 5;
+const INACTIVE_CLASS = 'ad-form--disabled';
 
+
+const INTERACTIVE_TAGS = [
+  'select',
+  'input',
+  'button',
+  'a',
+]
 
 const HOUSE_TYPE_MIN_PRICE = {
   flat: 1000,
@@ -48,4 +56,30 @@ const updateAddress = ({ lat, lng }) => {
   fieldAddress.value = `${lat.toFixed(DEGREE)}, ${lng.toFixed(DEGREE)}`;
 };
 
-export { initFormListeners, form, mapFilters, updateAddress };
+const makesFormsInactive = () => {
+  [form, mapFilters].forEach((el) => {
+    el.classList.add(INACTIVE_CLASS);
+    INTERACTIVE_TAGS.forEach((tag) => {
+      const children = el.querySelectorAll(tag);
+
+      children.forEach((elem) => {
+        elem.disabled = true;
+      });
+    })
+  })
+}
+
+const makesFormsActive = () => {
+  [form, mapFilters].forEach((el) => {
+    el.classList.remove(INACTIVE_CLASS);
+    INTERACTIVE_TAGS.forEach((tag) => {
+      const children = el.querySelectorAll(tag);
+
+      children.forEach((elem) => {
+        elem.disabled = false;
+      });
+    })
+  })
+};
+
+export { initFormListeners, makesFormsActive as makesFormActive, makesFormsInactive as makesFormInactive, form, mapFilters, updateAddress };
